Bug 1533612 - Change to unsigned int to avoid integer overflow. r=jya
mProcessedQueueLength is int32_t (wrapped around an Atomic) but it is updated from the method FramesToUsecs() which returns int64_t (wrapped around a CheckedInt) thus it is changed to unsigned int to avoid integer overflow.

Differential Revision: https://phabricator.services.mozilla.com/D22949

UltraBlame original commit: 814b2ae9ef66813cffc762db0353db2d0744fd81

+++ dom/media/mediasink/AudioSink.h
@@ -132,7 +132,7 @@ class AudioSink : private AudioStream::DataSource {
   UniquePtr<AudioConverter> mConverter;
   MediaQueue<AudioData> mProcessedQueue;
   
-  Atomic<int32_t> mProcessedQueueLength;
+  Atomic<uint64_t> mProcessedQueueLength;
   MediaEventListener mAudioQueueListener;
   MediaEventListener mAudioQueueFinishListener;
   MediaEventListener mProcessedQueueListener;
